LayoutTests/inspector/profiler/memory-instrumentation-cached-images-expected.txt

This test checks that CachedImages size reported by the memory agent includes images decoded size. Bug 93366


didReceiveMemorySnapshot
PASS: MemoryCache.Image size is OK.
